Pileup-based read error corrector
